perfil e correções

+++ bib/views.py
@@ -1,4 +1,6 @@
 from django.shortcuts import render, redirect
+
+from customauth.forms import UserChangeForm
 from .models import *
 from .forms import *
 from django.contrib.auth.decorators import login_required
@@ -11,7 +13,7 @@ def index(request):
 
 @login_required(login_url='auth.login')
 def index_book(request):   
-    books = Book.objects.all()
+    books = Book.objects.filter(user=request.user)    
     return render(request, 'books/index.html', { 'books': books })
 
 @login_required(login_url='auth.login')
@@ -27,8 +29,8 @@ def create_book(request):
 
 @login_required(login_url='auth.login')
 def show_book(request, id, pg):
-    book = Book.objects.get(pk=id)
-    if str(book.user) == str(request.user.email):
+    book = Book.objects.get(pk=id)    
+    if book.user.id == request.user.id:
         pdfFileObj = open(str(book.path), 'rb')
         pdfReader = PyPDF2.PdfFileReader(pdfFileObj)    
         count = pdfReader.numPages
@@ -52,17 +54,24 @@ def update_book(request):
 @login_required(login_url='auth.login')
 def delete_book(request, id):
     book = Book.objects.get(pk=id)
-    os.remove(str(book.path)) 
-    book.delete()
+    if book.user.id == request.user.id:
+        os.remove(str(book.path)) 
+        book.delete()
     return redirect('book.index')
 
 @login_required(login_url='auth.login')
-def config(request):   
-    return render(request, 'user/config.html')
+def profile(request, pk):   
+    user = MyUser.objects.get(pk=pk)
+    if user.id == request.user.id:
+        form = UserChangeForm(request.POST or None, instance=user)
+        if form.is_valid():
+            form.instance.user = request.user
+            form.save()
+            return redirect('./' + str(request.user.id))
+        
+        return render(request, 'user/profile.html', {"form":form})
 
-@login_required(login_url='auth.login')
-def profile(request):   
-    return render(request, 'user/profile.html')
+    return redirect('home')
 
 @login_required(login_url='auth.login')
 def dyslexic(request):
